ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "AI microservice for Viet Adventure Connect using FastAPI, LangChain/LangGraph"
authors = [
  { name = "Viet Adventure Connect" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "orjson>=3.10",
  "loguru>=0.7",
  "tenacity>=9.0",
  "langchain>=0.2",
  "langgraph>=0.2",
  "duckduckgo-search>=6.2",
  "google-generativeai>=0.7",
  "langchain-google-genai>=2.0",
  "tavily-python>=0.5",
  "firecrawl-py>=1.4",
  "pymongo>=4.15.5",
  "sseclient-py>=1.8.0",
  "requests>=2.32.5",
  "motor>=3.7.1",
  "langchain-core>=0.3.80",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
